Fix Setup API path
