+++ src/common/constants.js
@@ -7,5 +7,5 @@ const Bottle = require('bottlejs');
 
 exports = module.exports = {
   BOTTLE_CONTAINER_NAME: 'default',
-  BOTTLE_CONTAINER: Bottle.pop(exports.BOTTLE_CONTAINER_NAME),
+  BOTTLE_CONTAINER: Bottle.pop(exports.BOTTLE_CONTAINER_NAME)
 };
